修正B2CreateEdge的参数bug SoraPhysicalObject添加boundingBox绘制支持 SoraPhysicalObject添加getFixture, getBoundingBox SoraPhysicalObject添加RenderablePhysicalObject静态辅助函数 修正S偶然FreetypeFont::getGlyphByChar遇到不能识别的字符可能溢出的bug 修正SoraCore renderBox和renderRect颜色数据类型的bug SoraPath添加SoraPathImplOSX 修正SoraModifierList的repeat bug 修正一些数据类型转换错误 SoraOSX下开始针对app封装处理，添加CocoaApp项目 根目录添加Tutorials文件夹，用于存放教程项目，添加3个基础教程项目

+++ Sora/Defaults/SoraTimer_OSX.h
@@ -65,8 +65,8 @@ namespace sora {
 				if(oldtime != time) {
 					oldtime = time;
 					time = mach_absolute_time();
-					fdt = dtToMil(time-oldtime) / (float32)1000 * fTimeScale;
-					fps = 1.f / (dtToMil(time-oldtime)/(float32)1000);;
+					fdt = dtToMil(time-oldtime) / 1000.f * fTimeScale;
+					fps = 1.f / (dtToMil(time-oldtime)/1000.f);;
 				} else {
 					time = mach_absolute_time();
 				}
@@ -89,10 +89,10 @@ namespace sora {
 			time = mach_absolute_time();
 			
 			if(oldtime <= time - fpsInterval && oldtime >= time - (fpsInterval*2)) {
-				fdt = dtToMil(fpsInterval) / (float32)1000 * fTimeScale;
+				fdt = dtToMil(fpsInterval) / 1000.f * fTimeScale;
 				oldtime += fpsInterval;
 			} else {
-				fdt = dtToMil(time-oldtime) / (float32)1000 * fTimeScale;
+				fdt = dtToMil(time-oldtime) / 1000.f * fTimeScale;
 				oldtime = time;
 			}
 			if(fdt >= 1.f)
